moved root command for easier building
